+++ src/parser/classes/DDBClass.js
@@ -1346,7 +1346,7 @@ export default class DDBClass {
         title: this.is2014 ? "Ki Points" : "Focus Points",
         icon: null,
       };
-      utils.arrayRange(18, 1, 2).forEach((i) => {
+      utils.arrayRange(19, 1, 2).forEach((i) => {
         ki.configuration.scale[i] = {
           value: i,
         };